Updated documentation. Removed duplicated servlet code. Minor code clean up
