+++ frameworks/Java/servlet/README.md
@@ -1,29 +1,53 @@
-#Servlet Benchmarking Test
+# Servlet Benchmarking Test
 
 This is the Java Servlet portion of a [benchmarking test suite](../) comparing a variety of web development platforms.
 
-### JSON Encoding Test
-For raw Servlets there is no broad consensus on JSON encoding so we have selected the fastest available JSON encoder for Java: [Jackson](http://wiki.fasterxml.com/JacksonHome).
+### Plaintext and JSON
 
+* [Plaintext test source](src/main/java/hello/PlaintextServlet.java)
 * [JSON test source](src/main/java/hello/JsonServlet.java)
 
-### Data-Store/Database Mapping Test
+### `MySQL` implementation
+
 * [DB test source](src/main/java/hello/DbPoolServlet.java)
+* [Queries test source](src/main/java/hello/DbPoolServlet.java)
+* [Updates test source](src/main/java/hello/UpdateServlet.java) - using `batch updates`
+* [Fortune test source](src/main/java/hello/FortunesServlet.java)
+
+### `PostgreSQL` implementation
+
+DB, Queries and Fortune use the same implementation as MySQL.
+
+* [Updates test source](src/main/java/hello/PostgresUpdateServlet.java) - **not** using `batch updates` due to transaction deadlocks
+* [Cache test source](src/main/java/hello/Cache2kPostgresServlet.java)
 
 ## Infrastructure Software Versions
+
 The tests were run with:
 
-* [Java OpenJDK 1.7.0_09](http://openjdk.java.net/)
-* [Resin 4.0.34](http://www.caucho.com/)
-* [Jackson 2.3.0](http://wiki.fasterxml.com/JacksonHome)
-* [MySQL 5.5.29](https://dev.mysql.com/)
+* [Java Oracle JDK 1.8.0](http://openjdk.java.net/)
+* [Resin 4.0.53](http://www.caucho.com/)
+* [Jackson 2.8.9](http://wiki.fasterxml.com/JacksonHome)
+* [MySQL 5.7](https://dev.mysql.com/)
+* [Postgres 9.3](http://www.postgresql.org/)
+* [cache2k 1.0.0.CR4](https://cache2k.org/)
+
+Please confirm the versions data with the latest install scripts of TFB project.
 
 ## Test URLs
-### JSON Encoding Test
 
-http://localhost:8080/servlet/json
+### Default maven profile
 
-### Data-Store/Database Mapping Test
+ * Plaintext - `http://localhost:8080/servlet/plaintext`
+ * JSON - `http://localhost:8080/servlet/json`
 
-http://localhost:8080/servlet/db?queries=5
+### `mysql` and `postgresql` Maven profiles
 
+ * DB - `http://localhost:8080/servlet/db`
+ * Queries - `http://localhost:8080/servlet/db?queries=`
+ * Updates - `http://localhost:8080/servlet/update?queries=`
+ * Fortune - `http://localhost:8080/servlet/fortunes`
+
+### `postgresql` Maven profile
+ 
+ * Cache - `http://localhost:8080/servlet/cached-worlds`
