+++ frameworks/Java/servlet/benchmark_config.json
@@ -44,11 +44,11 @@
     },
     "postgres-raw": {
       "setup_file": "setup_postgresql",
-      "db_url": "/servlet/postgres",
-      "query_url": "/servlet/postgres?queries=",
+      "db_url": "/servlet/db",
+      "query_url": "/servlet/db?queries=",
       "cached_query_url": "/servlet/cached-worlds?queries=",
-      "fortune_url": "/servlet/postgres-fortunes",
-      "update_url": "/servlet/postgres-update?queries=",
+      "fortune_url": "/servlet/fortunes",
+      "update_url": "/servlet/update?queries=",
       "port": 8080,
       "approach": "Realistic",
       "classification": "Platform",
