+++ frameworks/Java/servlet/setup_postgresql.sh
@@ -2,7 +2,7 @@
 
 fw_depends java resin maven postgresql
 
-mvn clean install -P postgresql
+mvn clean compile war:war -P postgresql
 rm -rf $RESIN_HOME/webapps/*
 cp target/servlet.war $RESIN_HOME/webapps/
 resinctl start
