+++ frameworks/Java/servlet/src/main/resources/WEB-INF/postgresql/resin-web.xml
@@ -2,7 +2,7 @@
 
 	<database jndi-name='jdbc/hello_world'>
 		<driver>
-			<type>org.postgresql.Driver</type>
+			<type>org.postgresql.Driver</type><!-- TODO change this with the PooledConnection DS -->
 			<url>jdbc:postgresql://TFB-database:5432/hello_world</url>
 			<user>benchmarkdbuser</user>
 			<password>benchmarkdbpass</password>
